Banking Project 2 - Loan Defaulter Classification
This is a Banking Domain Project. The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed. 

The classification goal is to predict if the client will subscribe (yes/no) a term deposit . There are many categorical and numerical features . The folder  contains all plots , graphs and numerical calculations done in this project . People  who wants to learn about the practical  usage of Data preprocessing, EDA, pivot tables  , different classification models in Machine Learning along with diff tuning techniques can refer to this project